requirements.txt
certifi==2022.12.7
charset-normalizer==3.1.0
contourpy==1.0.7
cycler==0.11.0
dill==0.3.6
filelock==3.12.0
fonttools==4.39.3
h5py==3.8.0
idna==3.4
Jinja2==3.1.2
kiwisolver==1.4.4
loguru==0.7.0
MarkupSafe==2.1.2
matplotlib==3.7.1
mpmath==1.3.0
networkx==3.1
numpy==1.24.2
opencv-python-headless==4.7.0.72
packaging==23.1
pandas==2.0.0
Pillow==9.5.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2023.3
requests==2.28.2
six==1.16.0
sympy==1.11.1
torch==2.0.0
torch-snippets==0.224
torch-summary==1.4.5
torchmetrics==0.11.4
torchvision==0.15.1
tqdm==4.65.0
typing_extensions==4.5.0
tzdata==2023.3
urllib3==1.26.15
